+++ src/StatsdExporterClient.php
@@ -34,7 +34,7 @@ class StatsdExporterClient implements StatsdExporterClientInterface
     }
 
 
-    public function count(string $key, int $value, array $tags = []): StatsdExporterClientInterface
+    public function count(string $key, float $value, array $tags = []): StatsdExporterClientInterface
     {
         $this->statsdClient()->count($this->prepareKey($key), $value, $sampleRate = 1, $this->prepareTags($tags));
         return $this;
@@ -63,14 +63,14 @@ class StatsdExporterClient implements StatsdExporterClientInterface
     }
 
 
-    public function gauge(string $key, int $value, array $tags = []): StatsdExporterClientInterface
+    public function gauge(string $key, float $value, array $tags = []): StatsdExporterClientInterface
     {
         $this->statsdClient()->gauge($this->prepareKey($key), $value, $this->prepareTags($tags));
         return $this;
     }
 
 
-    public function summary(string $key, int $value, array $tags = []): StatsdExporterClientInterface
+    public function summary(string $key, float $value, array $tags = []): StatsdExporterClientInterface
     {
         $this->statsdClient()->timing($this->prepareKey($key), $value, $sampleRate = 1, $this->prepareTags($tags));
         return $this;
